requirements.txt
chromadb
sentence-transformers
openai
langchain
python-dotenv
pytest
